Migrate auth mixin to TS
