+++ src/data/ws-user.ts
@@ -1,20 +1,14 @@
 import {
-  createCollection,
   getUser,
   Connection,
+  getCollection,
 } from "home-assistant-js-websocket";
 import { User } from "../types";
+
+export const userCollection = (conn: Connection) =>
+  getCollection(conn, "_usr", () => getUser(conn) as Promise<User>, undefined);
 
 export const subscribeUser = (
   conn: Connection,
   onChange: (user: User) => void
-) =>
-  createCollection<User>(
-    "_usr",
-    // the getUser command is mistyped in current verrsion of HAWS.
-    // Fixed in 3.2.5
-    () => (getUser(conn) as unknown) as Promise<User>,
-    undefined,
-    conn,
-    onChange
-  );
+) => userCollection(conn).subscribe(onChange);
